+++ includes/class-plugin.php
@@ -62,6 +62,8 @@ class Plugin {
 			add_action( 'admin_menu', array( $admin_hooks, 'register_menu' ) );
 			add_action( 'admin_notices', array( $admin_hooks, 'render_notices' ) );
 			add_action( 'admin_enqueue_scripts', array( $admin_hooks, 'enqueue_assets' ) );
+			add_action( 'admin_post_tri_trash_restore', array( $admin_hooks, 'handle_trash_restore' ) );
+			add_action( 'admin_post_tri_trash_purge', array( $admin_hooks, 'handle_trash_purge' ) );
 		}
 	}